[FEATURE] Allow overwriting root paths
By adding the key 10 to layout-, partial- and templateRootPaths for the FLUIDTEMPLATE
it is now possible to outsource the files to another folder e.g. a template provider.
This is a functionality provided by most of the extensions in TER and should also be a
functionality provided by an extension created with EXT:mask_export